producer/requirements.txt
kafka-python
confluent-kafka[avro]
mmh3